+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/features/auth/redux/AuthRequests.kt
@@ -1,5 +1,6 @@
-package io.xorum.codeforceswatcher.features.auth
+package io.xorum.codeforceswatcher.features.auth.redux
 
+import io.xorum.codeforceswatcher.features.auth.UserAccount
 import io.xorum.codeforceswatcher.network.responses.backend.Response
 import io.xorum.codeforceswatcher.redux.*
 import io.xorum.codeforceswatcher.util.settings
