+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/features/auth/redux/AuthState.kt
@@ -1,4 +1,4 @@
-package io.xorum.codeforceswatcher.features.auth
+package io.xorum.codeforceswatcher.features.auth.redux
 
 import tw.geothings.rekotlin.StateType
 
@@ -9,5 +9,5 @@ data class AuthState(
 ) : StateType {
 
     enum class Status { IDLE, PENDING, DONE }
-    enum class Stage { NOT_SIGNED_IN, SIGNED_IN, VERIFIED}
+    enum class Stage { NOT_SIGNED_IN, SIGNED_IN, VERIFIED }
 }
