+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/features/users/redux/reducers/UsersReducer.kt
@@ -1,6 +1,6 @@
 package io.xorum.codeforceswatcher.features.users.redux.reducers
 
-import io.xorum.codeforceswatcher.features.auth.AuthRequests
+import io.xorum.codeforceswatcher.features.auth.redux.AuthRequests
 import io.xorum.codeforceswatcher.features.users.redux.actions.UsersActions
 import io.xorum.codeforceswatcher.features.users.redux.requests.UsersRequests
 import io.xorum.codeforceswatcher.features.users.redux.states.UsersState
