+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/redux/reducers/AppReducer.kt
@@ -1,6 +1,6 @@
 package io.xorum.codeforceswatcher.redux.reducers
 
-import io.xorum.codeforceswatcher.features.auth.authReducer
+import io.xorum.codeforceswatcher.features.auth.redux.authReducer
 import io.xorum.codeforceswatcher.features.news.redux.reducers.newsReducer
 import io.xorum.codeforceswatcher.features.users.redux.reducers.usersReducer
 import io.xorum.codeforceswatcher.features.problems.redux.reducers.problemsReducer
