+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/redux/states/AppState.kt
@@ -1,6 +1,6 @@
 package io.xorum.codeforceswatcher.redux.states
 
-import io.xorum.codeforceswatcher.features.auth.AuthState
+import io.xorum.codeforceswatcher.features.auth.redux.AuthState
 import io.xorum.codeforceswatcher.features.news.redux.states.NewsState
 import io.xorum.codeforceswatcher.features.problems.redux.states.ProblemsState
 import io.xorum.codeforceswatcher.features.contests.redux.states.ContestsState
